requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.50.0
httpx==0.27.2
sqlalchemy[asyncio]==2.0.35
asyncpg
aiosqlite
alembic==1.13.3
pdfplumber==0.11.4
Pillow==10.4.0
openpyxl==3.1.5
rapidfuzz==3.9.7
python-multipart==0.0.12
boto3==1.35.0
